requirements.txt
asgiref==3.7.2
Django==4.2.4
psycopg2-binary==2.9.7
sqlparse==0.4.4
DateTime~=5.2
djangorestframework~=3.14.0